Update README.md
This project employs Convolutional Neural Networks (CNNs) to detect various fruit infections and diseases. We've curated a dataset with diverse images showcasing different diseases like Angular Leaf Spot, Anthracnose Fruit Rot, and Blossom Blight. By training the CNN model on this dataset, we can accurately identify diseases in input images. Simply clone the repository, train the model, and use it to detect diseases in fruit images. Experience the power of AI as it assists in disease identification and contributes to agricultural health monitoring!